+++ test/util/test_edit.py
@@ -6,12 +6,17 @@ import zarr
 from test.sampledata import create_highroc_dataset
 from xcube.util.dsio import rimraf
 from xcube.util.edit import edit_metadata
+from xcube.util.optimize import optimize_dataset
 
 TEST_CUBE = create_highroc_dataset()
 
 TEST_CUBE_ZARR = 'test.zarr'
 
+TEST_CUBE_ZARR_OPTIMIZED = 'test-optimized.zarr'
+
 TEST_CUBE_ZARR_EDIT = 'test_edited_meta.zarr'
+
+TEST_CUBE_ZARR_OPTIMIZED_EDIT = 'test_optimized_edited_meta.zarr'
 
 TEST_NEW_META = {'output_metadata': {'creator_name': 'Brockmann Consult GmbH with love',
                                      'creator_url': 'www.some_very_nice_url.com'},
@@ -27,6 +32,8 @@ class EditVariablePropsTest(unittest.TestCase):
         rimraf(TEST_CUBE_ZARR)
         rimraf(TEST_NEW_META_YML)
         rimraf(TEST_CUBE_ZARR_EDIT)
+        rimraf(TEST_CUBE_ZARR_OPTIMIZED)
+        rimraf(TEST_CUBE_ZARR_OPTIMIZED_EDIT)
         TEST_CUBE.to_zarr(TEST_CUBE_ZARR)
         with open(TEST_NEW_META_YML, 'w') as outfile:
             yaml.dump(TEST_NEW_META, outfile, default_flow_style=False)
@@ -35,10 +42,12 @@ class EditVariablePropsTest(unittest.TestCase):
         rimraf(TEST_CUBE_ZARR)
         rimraf(TEST_NEW_META_YML)
         rimraf(TEST_CUBE_ZARR_EDIT)
+        rimraf(TEST_CUBE_ZARR_OPTIMIZED)
+        rimraf(TEST_CUBE_ZARR_OPTIMIZED_EDIT)
 
-    def test_edit_metadata_partly(self):
+    def test_edit_metadata(self):
         edit_metadata(TEST_CUBE_ZARR, metadata_path=TEST_NEW_META_YML, in_place=False,
-                      output_path=TEST_CUBE_ZARR_EDIT)
+                      output_path=TEST_CUBE_ZARR_EDIT, monitor=print)
         ds1 = zarr.open(TEST_CUBE_ZARR)
         ds2 = zarr.open(TEST_CUBE_ZARR_EDIT)
         self.assertEqual(ds1.__len__(), ds2.__len__())
@@ -47,8 +56,8 @@ class EditVariablePropsTest(unittest.TestCase):
         self.assertNotIn('creator_name', ds1.attrs.keys())
         self.assertIn('creator_name', ds2.attrs.keys())
 
-    def test_edit_metadata_partly_in_place(self):
-        edit_metadata(TEST_CUBE_ZARR, metadata_path=TEST_NEW_META_YML, in_place=True)
+    def test_edit_metadata_in_place(self):
+        edit_metadata(TEST_CUBE_ZARR, metadata_path=TEST_NEW_META_YML, in_place=True, monitor=print)
         ds1 = zarr.open(TEST_CUBE_ZARR)
         self.assertEqual(36, ds1.__len__())
         self.assertEqual('14-APR-2017 10:27:50.183264', ds1.attrs.__getitem__('start_date'))
@@ -57,3 +66,17 @@ class EditVariablePropsTest(unittest.TestCase):
         self.assertIn('creator_url', ds1.attrs.keys())
         self.assertEqual('Brockmann Consult GmbH with love', ds1.attrs.__getitem__('creator_name'))
         self.assertEqual('www.some_very_nice_url.com', ds1.attrs.__getitem__('creator_url'))
+
+    def test_edit_zmetadata(self):
+        edit_metadata(TEST_CUBE_ZARR, metadata_path=TEST_NEW_META_YML, in_place=False,
+                      output_path=TEST_CUBE_ZARR_EDIT, monitor=print)
+        optimize_dataset(TEST_CUBE_ZARR, unchunk_coords=True, output_path=TEST_CUBE_ZARR_OPTIMIZED)
+        edit_metadata(TEST_CUBE_ZARR_OPTIMIZED, metadata_path=TEST_NEW_META_YML, in_place=False,
+                      output_path=TEST_CUBE_ZARR_OPTIMIZED_EDIT, monitor=print)
+        ds1 = zarr.open(TEST_CUBE_ZARR)
+        ds2 = zarr.convenience.open_consolidated(TEST_CUBE_ZARR_OPTIMIZED_EDIT)
+        self.assertEqual(ds1.__len__(), ds2.__len__())
+        self.assertEqual(ds1.attrs.__getitem__('start_date'), ds2.attrs.__getitem__('start_date'))
+        self.assertEqual('happiness', ds2['conc_chl'].attrs.__getitem__('units'))
+        self.assertNotIn('creator_name', ds1.attrs.keys())
+        self.assertIn('creator_name', ds2.attrs.keys())