+++ xcube/api/__init__.py
@@ -16,12 +16,13 @@ from .verify import assert_cube, verify_cube
 # noinspection PyUnresolvedReferences
 from ..util.chunk import chunk_dataset, get_empty_dataset_chunks
 # noinspection PyUnresolvedReferences
-from ..util.geom import clip_dataset_by_geometry, mask_dataset_by_geometry, convert_geometry
+from ..util.edit import edit_metadata
+# noinspection PyUnresolvedReferences
+from ..util.geom import clip_dataset_by_geometry, convert_geometry, mask_dataset_by_geometry
 # noinspection PyUnresolvedReferences
 from ..util.maskset import MaskSet
 # noinspection PyUnresolvedReferences
 from ..util.optimize import optimize_dataset
-from ..util.edit import edit_metadata
 # noinspection PyUnresolvedReferences
 from ..util.unchunk import unchunk_dataset
 # noinspection PyUnresolvedReferences