+++ xcube/util/edit.py
@@ -20,17 +20,20 @@
 # SOFTWARE.
 import os.path
 import shutil
-from typing import Type
 import warnings
+from typing import Callable, Type
+
 import zarr
 
 from xcube.util.config import load_configs
+from xcube.util.optimize import optimize_dataset
 
 
 def edit_metadata(input_path: str,
                   output_path: str = None,
                   metadata_path: str = None,
                   in_place: bool = False,
+                  monitor: Callable[..., None] = None,
                   exception_type: Type[Exception] = ValueError):
     """
     Edit the metadata of an xcube dataset.
@@ -45,6 +48,7 @@ def edit_metadata(input_path: str,
     :param metadata_path: Path to the metadata file, which will edit the existing metadata.
     :param in_place: Whether to modify the dataset in place.
            If False, a copy is made and *output_path* must be given.
+    :param monitor: A progress monitor.
     :param exception_type: Type of exception to be used on value errors.
     """
 
@@ -70,21 +74,31 @@ def edit_metadata(input_path: str,
 
     new_metadata = load_configs(metadata_path)
 
+    if monitor is None:
+        # noinspection PyUnusedLocal
+        def monitor(*args):
+            pass
     for element in new_metadata:
-        if os.path.exists(os.path.join(output_path, '.zmetadata')):
-            print('.zmetadata exists')
+        cube = zarr.open(output_path)
         if 'output_metadata' in element:
-            cube = zarr.open(output_path)
-            _edit_keyvalue_in_metadata(cube, new_metadata, element)
+            _edit_keyvalue_in_metadata(cube, new_metadata, element, monitor)
         else:
-            cube = zarr.open(output_path)
             if cube.__contains__(element):
-                _edit_keyvalue_in_metadata(cube[element], new_metadata, element)
+                _edit_keyvalue_in_metadata(cube[element], new_metadata, element, monitor)
             else:
-                warnings.warn(f'The variable {element} could not be found in the xcube dataset. '
+                warnings.warn(f'The variable "{element}" could not be found in the xcube dataset. '
                               f'Please check spelling of it.')
+        # the metadata attrs of a consolidated xcube dataset may not be changed
+        # (https://zarr.readthedocs.io/en/stable/api/convenience.html#zarr.convenience.consolidate_metadata)
+        # therefore after changing metadata the xcube dataset needs to be consolidated once more.
+        if os.path.exists(os.path.join(output_path, '.zmetadata')):
+            optimize_dataset(output_path, in_place=True)
 
 
-def _edit_keyvalue_in_metadata(cube, new_metadata, element):
+def _edit_keyvalue_in_metadata(cube, new_metadata, element, monitor):
     for key in new_metadata[element].keys():
         cube.attrs.update({key: new_metadata[element][key]})
+        if 'output_metadata' in element:
+            monitor(f'Updated "{key}" in the global attributes of the xcube dataset.')
+        else:
+            monitor(f'Updated "{key}" in the attributes of "{element}" in the xcube dataset.')